# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/process": {
      "post": {
        "operationId": "ProcessController_trigger",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "processId": {
                    "enum": [
                      "1",
                      "2",
                      "3"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "Process"
        ]
      }
    },
    "/people": {
      "get": {
        "operationId": "PeopleController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "people"
        ]
      }
    },
    "/people/{id}": {
      "get": {
        "operationId": "PeopleController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "people"
        ]
      }
    },
    "/people/{id}/friends": {
      "get": {
        "operationId": "PeopleController_friends",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "people"
        ]
      }
    },
    "/accounts": {
      "get": {
        "operationId": "AccountsController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccount"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/person/{id}": {
      "get": {
        "operationId": "AccountsController_listByPerson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccount"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{iban}": {
      "get": {
        "operationId": "AccountsController_get",
        "parameters": [
          {
            "name": "iban",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccount"
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
    "/metrics/people/{id}/net-worth": {
      "get": {
        "operationId": "MetricsController_netWorth",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorth"
                }
              }
            }
          }
        },
        "tags": [
          "metrics"
        ]
      }
    },
    "/metrics/people/{id}/borrowable": {
      "get": {
        "operationId": "MetricsController_borrowable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Borrowable"
                }
              }
            }
          }
        },
        "tags": [
          "metrics"
        ]
      }
    },
    "/transactions": {
      "get": {
        "operationId": "TransactionsController_list",
        "parameters": [
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO date",
            "schema": {}
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO date",
            "schema": {}
          },
          {
            "name": "iban",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Transaction ID (uuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        },
        "tags": [
          "transactions"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "schemas": {
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "maxBorrowableCents": {
            "type": "number",
            "nullable": true
          },
          "netWorthCents": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "BankAccount": {
        "type": "object",
        "properties": {
          "iban": {
            "type": "string"
          },
          "balanceCents": {
            "type": "number"
          },
          "personId": {
            "type": "number"
          }
        },
        "required": [
          "iban",
          "balanceCents",
          "personId"
        ]
      },
      "NetWorth": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "number"
          },
          "netWorthCents": {
            "type": "number"
          }
        },
        "required": [
          "personId",
          "netWorthCents"
        ]
      },
      "Borrowable": {
        "type": "object",
        "properties": {
          "personId": {
            "type": "number"
          },
          "borrowableCents": {
            "type": "number"
          }
        },
        "required": [
          "personId",
          "borrowableCents"
        ]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountIban": {
            "type": "string"
          },
          "counterpartyIban": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "direction": {
            "type": "string",
            "enum": [
              "debit",
              "credit"
            ]
          },
          "loadedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "accountIban",
          "counterpartyIban",
          "amount",
          "direction",
          "loadedAt"
        ]
      }
    }
  }
}